separate py3 runtime tests from src tree + actually run said tests in CI
