+++ .gitignore
@@ -3,7 +3,7 @@ target/
 # ... but not code generation targets
 !tool/src/org/antlr/v4/codegen/target/
 
-# Node.js (npm and typings) cached dependencies 
+# Node.js (npm and typings) cached dependencies
 node_modules/
 typings/
 
@@ -98,3 +98,6 @@ xcuserdata
 javac-services.0.log
 javac-services.0.log.lck
 test/
+
+# Don't ignore python tests
+!runtime/Python3/test/
