+++ runtime/Python3/src/antlr4/InputStream.py
@@ -3,7 +3,6 @@
 # Use of this file is governed by the BSD 3-clause license that
 # can be found in the LICENSE.txt file in the project root.
 #
-import unittest
 
 
 #
@@ -85,20 +84,3 @@ class InputStream (object):
 
     def __str__(self):
         return self.strdata
-
-
-class TestInputStream(unittest.TestCase):
-
-    def testStream(self):
-        stream = InputStream("abcde")
-        self.assertEqual(0, stream.index)
-        self.assertEqual(5, stream.size)
-        self.assertEqual(ord("a"), stream.LA(1))
-        stream.consume()
-        self.assertEqual(1, stream.index)
-        stream.seek(5)
-        self.assertEqual(Token.EOF, stream.LA(1))
-        self.assertEqual("bcd", stream.getText(1, 3))
-        stream.reset()
-        self.assertEqual(0, stream.index)
-
