+++ runtime/Python3/src/antlr4/Recognizer.py
@@ -144,17 +144,3 @@ class Recognizer(object):
         self._stateNumber = atnState
 
 del RecognitionException
-
-import unittest
-class Test(unittest.TestCase):
-
-    def testVersion(self):
-        major, minor = Recognizer().extractVersion("1.2")
-        self.assertEqual("1", major)
-        self.assertEqual("2", minor)
-        major, minor = Recognizer().extractVersion("1.2.3")
-        self.assertEqual("1", major)
-        self.assertEqual("2", minor)
-        major, minor = Recognizer().extractVersion("1.2-snapshot")
-        self.assertEqual("1", major)
-        self.assertEqual("2", minor)
