+++ runtime/Python3/test/run.py
@@ -4,5 +4,9 @@ src_path = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__
 sys.path.insert(0,src_path)
 from TestTokenStreamRewriter import TestTokenStreamRewriter
 from xpathtest import XPathTest
+from TestFileStream import TestFileStream
+from TestInputStream import TestInputStream
+from TestIntervalSet import TestIntervalSet
+from TestRecognizer import TestRecognizer
 import unittest
 unittest.main()
